Extract AuditMember from model_soup
