+++ grouper/fe/handlers/audits_complete.py
@@ -2,8 +2,9 @@ from grouper.audit import get_audits
 from grouper.constants import PERMISSION_AUDITOR
 from grouper.email_util import cancel_async_emails
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import Audit, AUDIT_STATUS_CHOICES
+from grouper.model_soup import Audit
 from grouper.models.audit_log import AuditLog, AuditLogCategory
+from grouper.models.audit_member import AUDIT_STATUS_CHOICES
 from grouper.user_permissions import user_has_permission
